+++ holisticai/bias/mitigation/postprocessing/__init__.py
@@ -1,6 +1,22 @@
 # imports
 from .calibrated_eq_odds_postprocessing import CalibratedEqualizedOdds
 from .eq_odds_postprocessing import EqualizedOdds
+from .lp_debiaser.binary_balancer.transformer import LPDebiaserBinary
+from .lp_debiaser.multiclass_balancer.transformer import LPDebiaserMulticlass
+from .ml_debiaser.transformer import MLDebiaser
+from .plugin_estimator_and_recalibration.transformer import (
+    PluginEstimationAndCalibration,
+)
 from .reject_option_classification import RejectOptionClassification
+from .wasserstein_barycenters.transformer import WassersteinBarycenter
 
-__all__ = ["CalibratedEqualizedOdds", "EqualizedOdds", "RejectOptionClassification"]
+__all__ = [
+    "CalibratedEqualizedOdds",
+    "EqualizedOdds",
+    "RejectOptionClassification",
+    "WassersteinBarycenter",
+    "PluginEstimationAndCalibration",
+    "MLDebiaser",
+    "LPDebiaserBinary",
+    "LPDebiaserMulticlass",
+]