+++ holisticai/bias/mitigation/preprocessing/__init__.py
@@ -1,5 +1,12 @@
 # Imports
+from .correlation_remover import CorrelationRemover
+from .fairlet_clustering.transformer import FairletClusteringPreprocessing
 from .learning_fair_representation import LearningFairRepresentation
 from .reweighing import Reweighing
 
-__all__ = ["LearningFairRepresentation", "Reweighing"]
+__all__ = [
+    "LearningFairRepresentation",
+    "Reweighing",
+    "CorrelationRemover",
+    "FairletClusteringPreprocessing",
+]