+++ tests/bias/mitigation/reproducibility/test_eqodds.py
@@ -9,10 +9,7 @@ from sklearn.preprocessing import StandardScaler
 from holisticai.bias.metrics import classification_bias_metrics
 from holisticai.bias.mitigation import EqualizedOdds
 from holisticai.pipeline import Pipeline
-from tests.testing_utils._tests_data_utils import (
-    check_results,
-    load_preprocessed_adult_v2,
-)
+from tests.testing_utils._tests_utils import check_results, load_preprocessed_adult_v2
 
 seed = 42
 train_data, test_data = load_preprocessed_adult_v2()