+++ tests/bias/mitigation/reproducibility/test_learning-fair-representation.py
@@ -1,7 +1,6 @@
-import os
 import sys
 
-sys.path.append(os.getcwd())
+sys.path = ["./"] + sys.path
 
 import warnings
 
@@ -12,7 +11,7 @@ from sklearn.preprocessing import StandardScaler
 from holisticai.bias.metrics import classification_bias_metrics
 from holisticai.bias.mitigation import LearningFairRepresentation
 from holisticai.pipeline import Pipeline
-from tests.testing_utils._tests_data_utils import check_results, load_preprocessed_adult
+from tests.testing_utils._tests_utils import check_results, load_preprocessed_adult
 
 warnings.filterwarnings("ignore")
 
@@ -26,7 +25,7 @@ def running_without_pipeline():
     scaler = StandardScaler()
     Xt = scaler.fit_transform(X)
 
-    prep = LearningFairRepresentation(seed=seed)
+    prep = LearningFairRepresentation(k=10, Ax=0.1, Ay=1.0, Az=2.0, seed=seed)
     Xt = prep.fit_transform(Xt, y, **fit_params)
 
     model = LogisticRegression()
@@ -40,12 +39,7 @@ def running_without_pipeline():
     }
     Xt = prep.transform(Xt, **transform_params)
     y_pred = model.predict(Xt)
-    df = classification_bias_metrics(
-        group_b.to_numpy().ravel(),
-        group_a.to_numpy().ravel(),
-        y_pred,
-        y.to_numpy().ravel(),
-    )
+    df = classification_bias_metrics(group_b, group_a, y_pred, y, metric_type="both")
     return df
 
 
@@ -53,7 +47,10 @@ def running_with_pipeline():
     pipeline = Pipeline(
         steps=[
             ("scaler", StandardScaler()),
-            ("bm_preprocessing", LearningFairRepresentation(seed=seed)),
+            (
+                "bm_preprocessing",
+                LearningFairRepresentation(k=10, Ax=0.1, Ay=1.0, Az=2.0, seed=seed),
+            ),
             ("estimator", LogisticRegression()),
         ]
     )
@@ -69,12 +66,7 @@ def running_with_pipeline():
         "bm__group_b": group_b,
     }
     y_pred = pipeline.predict(X, **predict_params)
-    df = classification_bias_metrics(
-        group_b.to_numpy().ravel(),
-        group_a.to_numpy().ravel(),
-        y_pred,
-        y.to_numpy().ravel(),
-    )
+    df = classification_bias_metrics(group_b, group_a, y_pred, y, metric_type="both")
     return df